- 计费：发布时算平台参考价落库（reference_price / price_detail / need_manual_quote），
  用于展示与飞手报价锚点；订单金额仍由 selectRider 锁定为 quotedAmount——
  订单侧采 main 的 MATCHING 草稿流程，createOrder 回归两参。
- 重量字段统一为 main 的 cargoWeightKg（BigDecimal），删除我方 Double weight；
  PriceCalculator / PriceDetailVO / PriceEstimateDto 同步改 BigDecimal。
- reward 语义调整为「用户意向价」，不参与计费（后续用于按意向价筛选）。
- 迁移：原 task/*.sql 是 MySQL 专用语法（SET @var + PREPARE 守卫），而 main 的
  ADR-0001 已开启测试环境 Flyway，H2 跑不了。重写为兼容 MySQL 8.x 与 H2 MySQL
  模式的一次性迁移（对齐 V3 既有约定：不用任一侧特有语法、无需存在性守卫）；
  重编号 V6-V11 避开与 main 的 V4 冲突，表名按实体对齐（rider_feed* → feed*），
  补建 bill_config / user_address / task 计费列；删除 FlywayConfig（main 已由
  spring-boot 标准配置接管）。

验证：mvnw verify → 189 测试 0 失败（含 TransportPriceCalculatorTest / MatchFlowIT /
TaskApplicationIT / OrderAmountGuardIT / OpenApiContractGate）；契约重导 114 操作。

+++ drone-backend/src/main/java/com/uav/task/pojo/dto/PriceEstimateDto.java
@@ -3,6 +3,7 @@ package com.uav.task.pojo.dto;
 import com.uav.server.enums.TaskType;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 /** 计价请求：与发布任务同口径，预览参考价用 */
@@ -10,7 +11,8 @@ import java.util.List;
 public class PriceEstimateDto {
     private TaskType taskType;
     private List<WaypointDto> waypoints;
-    private Double weight;
+    /** 吊运货物重量（kg），与 {@code TaskDto.cargoWeightKg} 同口径 */
+    private BigDecimal weight;
     /** 计划执行时间，'yyyy-MM-dd HH:mm:ss'，用于夜间附加费判断（可空） */
     private String taskTime;
 }
